+++ src/pages/droner/EditDroner.tsx
@@ -426,11 +426,13 @@ function EditDroner() {
     setEditIndex(0);
     setDronerDroneList(dronerDroneList);
   };
-  const removeDrone = (index: number) => {
-    const newData = dronerDroneList.filter(
-      (x) => x.modalDroneIndex !== index
-    );
-    setDronerDroneList(newData);
+  const removeDrone = async (id?: string) => {
+    try {
+      await DronerDroneDatasource.removeDronerDrone(id);
+      fetchDronerById();
+    } catch (e) {
+      console.log(e);
+    }
   };
   //#endregion
 
@@ -595,8 +597,6 @@ function EditDroner() {
   };
 
   const onFieldsChange = (field: any) => {
-    const currentStatus =
-      (field[0].name[0] === "status" && field[0].value) || status;
     const isHasError = form.getFieldsError().some(({ errors }) => {
       return errors.length > 0;
     });
@@ -609,6 +609,7 @@ function EditDroner() {
       checkPlantsOther,
       reason,
       idNo,
+      status: currentStatus,
       ...rest
     } = form.getFieldsValue();
     const reasonList = [];
@@ -902,9 +903,11 @@ function EditDroner() {
                   onChange={handleProvince}
                   defaultValue={address.provinceId}>
                   {province.map((item: any, index: any) => (
-                    <option key={index} value={item.provinceId}>
+                    <Select.Option
+                      key={index}
+                      value={item.provinceId}>
                       {item.provinceName}
-                    </option>
+                    </Select.Option>
                   ))}
                 </Select>
               </Form.Item>
@@ -938,9 +941,11 @@ function EditDroner() {
                   onChange={handleDistrict}
                   defaultValue={address.districtId}>
                   {district.map((item: any, index: any) => (
-                    <option key={index} value={item.districtId}>
+                    <Select.Option
+                      key={index}
+                      value={item.districtId}>
                       {item.districtName}
-                    </option>
+                    </Select.Option>
                   ))}
                 </Select>
               </Form.Item>
@@ -976,9 +981,11 @@ function EditDroner() {
                   onChange={handleSubDistrict}
                   defaultValue={address.subdistrictId}>
                   {subdistrict?.map((item: any, index: any) => (
-                    <option key={index} value={item.subdistrictId}>
+                    <Select.Option
+                      key={index}
+                      value={item.subdistrictId}>
                       {item.subdistrictName}
-                    </option>
+                    </Select.Option>
                   ))}
                 </Select>
               </Form.Item>
@@ -1061,7 +1068,7 @@ function EditDroner() {
                     option.children.includes(input)
                   }>
                   {location.map((item: any) => (
-                    <option
+                    <Select.Option
                       key={item.subdistrictId}
                       value={item.subdistrictId}>
                       {item.subdistrictName +
@@ -1069,7 +1076,7 @@ function EditDroner() {
                         item.districtName +
                         "/" +
                         item.provinceName}
-                    </option>
+                    </Select.Option>
                   ))}
                 </Select>
               </Form.Item>
@@ -1147,16 +1154,16 @@ function EditDroner() {
               }}
               rules={[
                 {
-                  validator: (_, value, callback) => {
+                  validator: (_, value) => {
                     const plantsOther =
                       form.getFieldValue("plantsOther");
 
                     if (value?.length < 1 && !plantsOther) {
-                      callback(
+                      return Promise.reject(
                         "กรุณาเลือกพืชที่เคยฉีดพ่นอย่างน้อย 1 อย่าง"
                       );
                     } else {
-                      callback();
+                      return Promise.resolve();
                     }
                   },
                 },
@@ -1193,6 +1200,7 @@ function EditDroner() {
                       );
 
                     const isDupTyping =
+                      splitValue &&
                       new Set(splitValue).size !== splitValue.length;
 
                     if (!!value && checkValidateComma(value)) {
@@ -1247,11 +1255,7 @@ function EditDroner() {
                                 name="checkReason"
                                 rules={[
                                   {
-                                    validator: (
-                                      _,
-                                      value,
-                                      callback
-                                    ) => {
+                                    validator: (_, value) => {
                                       const reasonField =
                                         form.getFieldValue("reason");
 
@@ -1416,7 +1420,7 @@ function EditDroner() {
                       <ActionButton
                         icon={<DeleteOutlined />}
                         color={color.Error}
-                        onClick={() => removeDrone(index + 1)}
+                        onClick={() => removeDrone(item.id)}
                       />
                     </div>
                   </div>