+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## [Version - 0]
+## Fix [v0.17.14] 2022-10-17
+- fixed code upload image Edit Add Farmer & Droner
+- fixed delete droner-drone
+## Fix [v0.17.13] 2022-10-17
+- fixed cant submit form issue AddDroner  & EditDroner
 ## Fix [v0.17.12] 2022-10-14
 - increase validate Tel number  Droner & Farmer (Add & Edit)
 - fixed Lat long not changing when change address